fix(lorawan): review nits - FCtrl printf cast + protocol-neutral RF hint
Adversarial multi-agent review (5 dimensions, 26 agents) of the
main..v8.3-dev diff: 21 findings raised -> 2 confirmed (both cosmetic
nits), 19 refuted. Parser bounds-math, loop-safety/dedup, the
g_mcClock*->g_clock* rename, and the portal validator all verified clean;
no functional bugs.

- printLoRaWAN: cast m.fctrl to (unsigned) for the %02X specifier
  (strict-conformance signedness; output was already correct for 0..255).
- CaptivePortal: the BW/SF/CR hint hard-coded "MeshCore:" but that block
  is now also shown under LoRaWAN - made the hint protocol-neutral.
- CLAUDE.md: record the review result.

Build green (pio run -e xiao_esp32s3): Flash 24.6%.

+++ src/MeshDecoderDebug.h
@@ -1030,7 +1030,7 @@ inline void printLoRaWAN(const uint8_t *buf, size_t len,
         Serial.printf("[%8lu ms][%s decoded] LoRaWAN %s DevAddr=0x%08lx FCtrl=0x%02X "
                       "FCnt=%u FPort=%d FRMlen=%u len=%u\n",
                       millis(), tag, loraWanMtypeName(m.mtype),
-                      (unsigned long)m.devAddr, m.fctrl, (unsigned)m.fcnt,
+                      (unsigned long)m.devAddr, (unsigned)m.fctrl, (unsigned)m.fcnt,
                       m.hasFport ? (int)m.fport : -1, (unsigned)m.frmPayloadLen, (unsigned)len);
     } else {
         Serial.printf("[%8lu ms][%s decoded] LoRaWAN %s (encrypted/opaque) len=%u\n",
